refactors reducer
