+++ project/src/store/movie-data/movie-data.js
@@ -1,5 +1,5 @@
-import {INITIAL_GENRE, AuthorizationStatus} from '../consts.js';
-import {ActionType} from './action.js';
+import { ActionType } from '../action';
+import { INITIAL_GENRE } from '../../consts';
 
 const initialState = {
   genre: INITIAL_GENRE,
@@ -7,12 +7,11 @@ const initialState = {
   promoMovie: {},
   favoriteMovies: [],
   reviews: [],
-  authorizationStatus: AuthorizationStatus.UNKNOWN,
   isFetching: false,
   error: '',
 };
 
-export const reducer = (state = initialState, action) => {
+export const movieData = (state = initialState, action) => {
   switch (action.type) {
     case ActionType.CHANGE_GENRE:
       return {
@@ -40,22 +39,6 @@ export const reducer = (state = initialState, action) => {
         ...state,
         favoriteMovies: action.payload,
         isFetching: true,
-      };
-    case ActionType.LOAD_REVIEWS:
-      return {
-        ...state,
-        reviews: action.payload,
-        isFetching: true,
-      };
-    case ActionType.REQUIRE_AUTHORIZATION:
-      return {
-        ...state,
-        authorizationStatus: action.payload,
-      };
-    case ActionType.LOGOUT:
-      return {
-        ...state,
-        authorizationStatus: AuthorizationStatus.NO_AUTH,
       };
     case ActionType.ERROR:
       return {
